Change PeerPage 'body' field to 'page_text'
Because this field needs to store rich text, we can't apply the standard
HTML input encoding that CiviCRM uses. The easiest way to skip the input
encoding is to choose a field name CiviCRM skips by default. These
fields are given in \CRM_Utils_API_HTMLInputCoder::getSkipFields() and
'page_text' is a skipped field (but 'body' is not).

Worth mentioning that I think this whole design of "skipped fields" is
very smelly from a security standpoint, but it would take more time to
fix (in core) than I have right now.